server/requirements.txt
openenv-core>=0.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.25.0
openai>=1.0.0
